Décalé le code de l'IA dans une classe séparé
voir IA.py tout le code a été décalée dedans.
